+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-06 09:18:52
-// Commit: docs(state): update readme and API contracts for state
+// Updated at: 2026-05-06 14:04:35
+// Commit: fix(state): fix memory leak by cancelling state subscription
 
 class stateComponentState {
-  final String commitRef = '921';
-  final String lastMessage = 'docs(state): update readme and API contracts for state';
-  final DateTime timestamp = DateTime.parse('2026-05-06T09:18:52+05:00');
+  final String commitRef = '937';
+  final String lastMessage = 'fix(state): fix memory leak by cancelling state subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-06T14:04:35+05:00');
 }